backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.109.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tomli==2.2.1
typing_extensions==4.13.2
uvicorn==0.27.1
